requirements.txt
torch
torchvision
transformers
numpy
matplotlib
scikit-learn
seaborn
tqdm
pandas
pillow
flask